Support for multiple status elements
XCCDF allows multiple <status> elements in benchmark / profile
elements. This commit changes old behavior that was supporting
only one <status> element.

Datetime timestamp added (specification says it's OPTIONAL but
in case of Tailoring file it is not).

This commit also unifies status and version objects and switches to
fixed structure for status element:

  status:
    - value: <status value>
      timestamp: <status timestamp>